refactor: Update HeroCards and CTAbutton components
The HeroCards component was modified to include the "use client" import statement, improving the performance of the page load by lazy loading images. Additionally, the CTAbutton component was updated to receive the currentUser prop, allowing for conditional rendering based on the user's authentication status. These changes enhance the overall user experience and optimize the functionality of the components.